requirements.txt
azure-storage-blob==12.20.0
Flask==2.1.2
tensorflow-cpu==2.16.1
numpy==1.26.4
Pillow==9.4.0
pytest==7.4.0